Add comprehensive test suite for FoldDB core functionality
- Implement test modules for Atom, FoldDB, Permissions, Schema, and SchemaMapper
- Add integration tests covering database operations, permissions, and schema transformations
- Create test utilities for database setup, cleanup, and consistent test environment
- Enhance test coverage for permission policies, trust distance, and field access
- Improve error handling and validation in test scenarios
- Add robust test cases for schema mapping, field management, and data versioning

+++ src/permissions/permission_manager.rs
@@ -5,16 +5,15 @@ pub struct PermissionManager {
 }
 
 impl PermissionManager {
-    pub fn has_read_permission(&self, pub_key: &str, permissions_policy: &PermissionsPolicy, trust_distance: u8) -> bool {
+    pub fn has_read_permission(&self, pub_key: &str, permissions_policy: &PermissionsPolicy, trust_distance: u32) -> bool {
         permissions_policy.read_policy >= trust_distance || 
         permissions_policy.explicit_read_policy.as_ref()
             .map_or(false, |counts| counts.counts_by_pub_key.get(pub_key).is_some())
     }
 
-    pub fn has_write_permission(&self, pub_key: &str, permissions_policy: &PermissionsPolicy, trust_distance: u8) -> bool {
+    pub fn has_write_permission(&self, pub_key: &str, permissions_policy: &PermissionsPolicy, trust_distance: u32) -> bool {
         permissions_policy.write_policy >= trust_distance || 
         permissions_policy.explicit_write_policy.as_ref()
             .map_or(false, |counts| counts.counts_by_pub_key.get(pub_key).is_some())
     }
 }
-
